Remove "contianed_in" as an annotation option
On 9/19/2014 in our weekly SemGen / VPR meeting decided to remove
"contained_in" as an annotation descriptor option and leave "part_of" as
the label describing the relationship between an entity and its parent.

In these changes I removed the combobox containing "contained_in" and
"part_of" and in its place inserted a label displaying "part_of".