Receive commands from JS
These changes complete the first step to receiving commands in JS. These
changes contain  and abstract class that will be used to receive
commends in java and send commands from javascript.

I also added an event, cwb-initialized, that's fired when the browser is
initialized. This enables the js code to listen for that event before it
reacts.