Add parent node
Submodels will also be parents, but they wont share the same behaviors
that the model nodes share. So, in these changes we move the parent
behavior to its own class so the model node and submodel node (or
component node, haven't created it yet, nor have I decided on a name)
can share the parent functionality.