Mock - Show submodel dependency network
When a submodel is clicked let java know (the mock). Java (in this case
the mock) will then tell the stage to visualize the submodel's
dependency network. Javascript does this by finding the parent model,
it's submodel, then adding children to that submodel.

The visualization is a little off. More changes to come.