Links pointing in the wrong direction
SemGen encodes links as inputs. So if node A has links A.links = ["B",
"C", "D"] then "B", "C" and "D" should have arrows pointing to A. I was
treating links as outputs. This changes fixes that issue and treats
links as inputs.

I hope to follow these changes with a change that updates the name of
the array from "links" to "inputs" for clarity. For now (for the demo),
I'll just keep the change simple.