Add hull to submodel nodes
The hull now needs to analyze all descendant nodes in order to position
itself appropriately, since children can have children (in the case of a
submodel).

Also, when the hull is collapsed we need to remove all descendant nodes
from the graph. This is for the case where you expand a model node into
submodel nodes, then expand a submodel node into a dependency network,
then collapse the parent model node. In that case, you need to hide all
descendants of the model node (submodels and their children), not just
the first descendants (submodels)