Add Search Mock
Add mock searching so we can test searching outside of SemGen. In doing
so I also updated the search result values that are passed from Java to
JS (one of the perks of using mock is that it allows you to test out the
interface so you can make it behave how you want it to before actually
implementing). Instead of converting a set to json we now pass a json
array to JS so it doesn't have to do any extra manipulation.