Highlight nodes on mouseover
When you mouseover a nodes these changes will highlight that node, its
links, and all of the nodes linked to it. Highlighting happens by
dimming out all non highlighted nodes. This way we can use CSS to style
the dim. For example, the model nodes do not dim, which we were able to
achieve in CSS. This was my attempt at keeping the graph agnostic of
certain behaviors so it can be used for other things in the future.