Reading and writing CellML files tests
This is the initial check in for the set of tests that will test the
end-end reading and writing cellml files. Unit testing the cellml reader
and writer is too expensive (we don't need this to be production
quality) so we're creating end-end tests to test everything at once in
one swoop.