Update the height and width of the graph
For some reason with the new framework we're loading before the height
and width of the window is set. This probably means we're loading the
graph before the window is completely loaded, but I'm not 100% sure.

Rather than spend time figuring it out, I implemented a change I've been
meaning to implement that solves our 0 height 0 width problem. Now,
every time we draw the graph we update the width and height of the graph
to the size of the window. So if you resize the window and update the
graph (add/remove a node) the graph will resize.