Load the stage from a temp dir
JxBrowser can't load files from within a jar. As a result, we need to
copy the files (i.e. the stage and all of it's js, css, and image
resources) to a temporary directory before they are loaded. I grabbed a
class off of stackoverflow to help with the copying, and the rest was
pretty simple.

This is in reference to #24.