Large node refactor
The need for visualizing additional nodes required a significant
refactor. Basically, on the Java side we now generate ids and created a
Link class for transffering information about links to JS, and on the JS
side we removed a lot of the uneeded logic that now lives in the Java
side.

To see examples of the new data structures see mock.js.

There are still a bunch of comments I'd like to add, but since time is
against me I'm checking in without them :/.