Move stage files to an external folder
I was trying to avoid having another external folder, but the trouble I
was having wasn't worth it. These changes move the stage to an external
folder (outside of the jar) and update SemGen to point to the external
folder to load the stage.

It would have been nice to keep the folder inside of the jar to decrease
the number of files and folders that sit along side of it. I don't love
this solution, but it works for this research project.