Finish End-End Tests
These changes finish end-end testing (I hope). There are two main
changes here:

1) Updating the collateral file. The original model file had things that
weren't getting written out to the test model file. For example, the
xmlns:dc attribute on the <model> node in the original file was not
written out to the test model file. There were several instances where
this type of thing was happening, so I updated the file to only contain
information that was written out after the file was processed. The
advantage of this is that we're now verifying that SemGen can properly
reads in model content that we expect, translate it to a SemSimModel
object, and write it back out to a file. The disadvantage of this
approach is that in practice SemGen handles files that have "extra"
attributes that aren't written out in by the CellMLWriter, and those
scenarios are not covered by this test. However, I think this is a
sufficient first step and gives us good test coverage on the things we
care about most.

2) I updated the test that reads, writes and compares the original model
file to the test model file. Interestingly enough, during testing I
discovered that SemGen writes out the <units> elements (and maybe
others) in a nondeterministic order. However, this test now takes care
of that by searching for elements with matching names and attributes.

I also deleted ModelCopyTest because it contained a stale test that we
don't need or want anymore.