Update transformer class that's used
These changes update the transformer class that was used because the old
transformer was causing invalid results. The old transformer was
improperly modifying the document tree leaving a tree that was ruined
and caused an invalid comparison. These changes use a transformer that
works as expected.

I'm not sure exactly what the transformer class is doing, but it's used
when modifying the document tree to ignore and normalize whitespace,
ignore comments, diff between text and cdata, and ignore attribute
order. Since using it we can ignore the things we want to ignore and
diff properly. Now that we're using a proper transformer we're seeing
things that were hidden by the old one.

The <connect> tags don't have attributes (just child elements), and
since the comparator matches elements based on tag name and attribute,
tags from the source document are compared against the wrong nodes in
the test document (since <connection> and <connection> look the same the
comparator says, "hey, these are matching elements", when it really
should look at it's children to find that out). We could ignore this, or
we could fix it in one of two ways: 1) Make sure the <connection>
elements that we write out are in the same order at the connection
elements that we read in, or 2) only compare <connection> elements if
the children match (significant amount of work).