Read/write annotations on custom physical properties for CellML models.
In making these changes I no longer allow users to annotate custom
entity terms using the part-of relation. This is because in the RDF
added to CellML models, there is currently no way to distinguish part-of
statements that link a custom entity to another entity within a
composite, and those that indicate that a custom entity is part-of some
reference term in general.